worked on reducers, almost finished
