+++ client/components/reducers/index.js
@@ -1,10 +1,12 @@
 import {configureStore} from '@reduxjs/toolkit'
-import swoopSlice from './swoopSlice.js'
+import userReducer from './userSlice'
+import itemReducer from './itemSlice';
 
 const store = configureStore({
     devTools: true,
     reducer: {
-        swoop: swoopSlice,
+        user: userReducer,
+        item: itemReducer
     }
 });
