+++ client/components/sidenav/Login.jsx
@@ -40,5 +40,5 @@ return (
   </div>
 )
 }
-
+// why is password's id like that?
 export default Login;
